+++ android/src/main/java/com/brentvatne/receiver/BecomingNoisyListener.kt
@@ -1,0 +1,12 @@
+package com.brentvatne.receiver
+
+interface BecomingNoisyListener {
+    companion object {
+        val NO_OP = object : BecomingNoisyListener {
+            override fun onAudioBecomingNoisy() {
+                // NO_OP
+            }
+        }
+    }
+    fun onAudioBecomingNoisy()
+}